Version 0.1.1 / Develop-ToMain

+++ src/main/java/com/example/demo/Dtos/LikeRequest.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 
 public class LikeRequest {
-    private long userId;
-    private long postId;
+    private Integer userId;
+    private Integer postId;
 }
