+++ src/main/java/com/example/demo/Dtos/LikeResponse.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 
 public class LikeResponse {
-    private long likeId;
+    private int likeId;
 }
