+++ src/main/java/com/example/demo/Models/Like.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class Like {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long likeId;
+    private int likeId;
     @ManyToOne
     @JoinColumn(name = "userId", nullable = false)
     private User user;
