+++ src/main/java/com/example/demo/Repositories/LikeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LikeRepository extends JpaRepository<Like, Long> {
+public interface LikeRepository extends JpaRepository<Like, Integer> {
     boolean existsByUserAndPost(User user, Post post);
     // El optional es porque puede ser o no vacio y encontrar el "me gusta
     //Optional<Like> findByUserAndPost(User user, Post post);
